Move borrow button click handler  (#147)
