+++ frontend/app/components/cart/CartSelectedDialog.tsx
@@ -1,19 +1,33 @@
 import { Button, Dialog, Stack } from '@mantine/core';
+import { useSubmit } from '@remix-run/react';
 import { useAtom } from 'jotai';
-import { cartAtom, selectedCartBooksAtom } from '~/stores/cartAtom';
+import { cartAtom, CartProps, selectedCartBooksAtom } from '~/stores/cartAtom';
 import { removeBooksFromCart } from '~/utils/cart';
+import { errorNotification } from '~/utils/notification';
 
-interface CartSelectedDialogProps {
-	handleBorrowButtonClick: () => void;
-}
-
-const CartSelectedDialog = ({
-	handleBorrowButtonClick,
-}: CartSelectedDialogProps) => {
-	const [selectedCartBook, setSelectedCartBook] = useAtom(
-		selectedCartBooksAtom,
-	);
+const CartSelectedDialog = () => {
+	// prettier-ignore
+	const [selectedCartBook, setSelectedCartBook] = useAtom(selectedCartBooksAtom);
 	const [cart, setCart] = useAtom(cartAtom);
+
+	const submit = useSubmit();
+
+	// 貸し出し数が蔵書数を超えていないかチェックする
+	const checkStock = (element: CartProps) => element.stock < element.volume;
+
+	const handleBorrowButtonClick = () => {
+		if (selectedCartBook.length > 0 && !selectedCartBook.some(checkStock)) {
+			submit(JSON.stringify({ selectedCartBook: selectedCartBook }), {
+				action: '/home/cart',
+				method: 'PATCH',
+				encType: 'application/json',
+			});
+			setCart(removeBooksFromCart(cart, selectedCartBook));
+			setSelectedCartBook([]);
+		} else {
+			errorNotification('在庫が足りません');
+		}
+	};
 
 	return (
 		<Dialog
